+++ EtLast.AdoNet/Mutators/WriteToMsSqlMutator.cs
@@ -31,7 +31,7 @@ public sealed class WriteToMsSqlMutator : AbstractMutator, IRowSink
     /// </summary>
     public int ForceWriteAfterNoDataMilliseconds { get; init; } = 5000;
 
-    private int _rowsWritten;
+    private long _rowsWritten;
     private DatabaseConnection _connection;
     private SqlBulkCopy _bulkCopy;
     private RowShadowReader _reader;
