+++ EtLast/Streams/NamedSink.cs
@@ -3,7 +3,7 @@
 public class NamedSink : NamedStream
 {
     public int SinkUid { get; }
-    public int RowsWritten { get; private set; }
+    public long RowsWritten { get; private set; }
 
     public NamedSink(string name, Stream stream, int ioCommandUid, IoCommandKind ioCommandKind, int sinkUid)
         : base(name, stream, ioCommandUid, ioCommandKind)
