+++ EtLast/Streams/SinkProvider/PartitionKeyGenerator.cs
@@ -1,3 +1,3 @@
 ﻿namespace FizzCode.EtLast;
 
-public delegate string PartitionKeyGenerator(IRow row, int index);
+public delegate string PartitionKeyGenerator(IRow row, long index);
